+++ facebook_auth/lib/flutter_facebook_auth.dart
@@ -19,14 +19,14 @@ class FacebookAuth {
   /// you can use FacebookAuth.instance or FacebookAuth.i
   static FacebookAuth get i => instance;
 
-  FacebookAuthPlatform _ = FacebookAuthPlatform.instance;
+  final FacebookAuthPlatform _authPlatform = FacebookAuthPlatform.instance;
 
   /// return a new instance of FacebookAuth for testing
   @visibleForTesting
   static FacebookAuth getInstance() => FacebookAuth._();
 
   /// if the user is logged return one instance of AccessToken
-  Future<AccessToken?> get accessToken => _.accessToken;
+  Future<AccessToken?> get accessToken => _authPlatform.accessToken;
 
   /// Express login logs people in with their Facebook account across devices and platform.
   /// If a person logs into your app on Android and then changes devices,
@@ -35,7 +35,7 @@ class FacebookAuth {
   /// This avoid creating duplicate accounts or failing to log in at all. To support the changes in Android 11,
   /// first add the following code to the queries element in your /app/manifest/AndroidManifest.xml file.
   /// For more info go to https://developers.facebook.com/docs/facebook-login/android
-  Future<LoginResult> expressLogin() => _.expressLogin();
+  Future<LoginResult> expressLogin() => _authPlatform.expressLogin();
 
   /// retrive the user information using the GraphAPI
   ///
@@ -49,7 +49,7 @@ class FacebookAuth {
   ///  "picture": {
   ///    "data": {
   ///      "height": 126,
-  ///      "url": "https://scontent.fuio21-1.fna.fbcdn.net/v/t1.30497-1/s200x200/84628273_176159830277856_972693363922829312_n.jpg",
+  ///      "url": "https://scontent.fuio21-1.fna.fbcdn.net/v/t1.30497-1/s200x200/8462.jpg",
   ///      "width": 200
   ///    }
   ///  },
@@ -61,10 +61,10 @@ class FacebookAuth {
   Future<Map<String, dynamic>> getUserData({
     String fields = "name,email,picture.width(200)",
   }) =>
-      _.getUserData(fields: fields);
+      _authPlatform.getUserData(fields: fields);
 
   /// Sign Out from Facebook
-  Future<void> logOut() => _.logOut();
+  Future<void> logOut() => _authPlatform.logOut();
 
   /// make a login request using the facebook SDK
   ///
@@ -76,16 +76,19 @@ class FacebookAuth {
     List<String> permissions = const ['email', 'public_profile'],
     LoginBehavior loginBehavior = LoginBehavior.dialogOnly,
   }) =>
-      _.login(permissions: permissions, loginBehavior: loginBehavior);
+      _authPlatform.login(
+        permissions: permissions,
+        loginBehavior: loginBehavior,
+      );
 
   /// call this method (ONLY FOR WEB) to initialize the facebook javascript sdk
-  void webInitialize({
+  Future<void> webInitialize({
     required String appId,
     required bool cookie,
     required bool xfbml,
     required String version,
   }) {
-    _.webInitialize(
+    return _authPlatform.webInitialize(
       appId: appId,
       cookie: cookie,
       xfbml: xfbml,
@@ -96,18 +99,19 @@ class FacebookAuth {
   /// returns one instance of FacebookPermission with the granted and declined permissions
   ///
   /// It could be null if you exceed the request limit
-  Future<FacebookPermissions?> get permissions => _.permissions;
+  Future<FacebookPermissions?> get permissions => _authPlatform.permissions;
 
   /// use this to know if the facebook sdk was initializated on Web
   /// on Android and iOS is always true
-  bool get isWebSdkInitialized => _.isWebSdkInitialized;
+  bool get isWebSdkInitialized => _authPlatform.isWebSdkInitialized;
 
   /// ONLY FOR iOS: enable or disable AutoLogAppEvents
   Future<void> autoLogAppEventsEnabled(bool enabled) {
-    return _.autoLogAppEventsEnabled(enabled);
+    return _authPlatform.autoLogAppEventsEnabled(enabled);
   }
 
   /// ONLY FOR iOS: check if AutoLogAppEvents are enabled
 
-  Future<bool> get isAutoLogAppEventsEnabled => _.isAutoLogAppEventsEnabled;
+  Future<bool> get isAutoLogAppEventsEnabled =>
+      _authPlatform.isAutoLogAppEventsEnabled;
 }